requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
soupsieve==2.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
